+++ src/onegov.election_day/onegov/election_day/tests/forms/test_election_form.py
@@ -26,9 +26,9 @@ def test_election_form_domains():
     ]
 
 
-def test_election_form_translations():
+def test_election_form_translations(session):
     form = ElectionForm()
-    form.request = DummyRequest()
+    form.request = DummyRequest(session=session)
     form.request.default_locale = 'de_CH'
     form.on_request()
     assert isinstance(form.election_de.validators[0], InputRequired)
@@ -37,7 +37,7 @@ def test_election_form_translations():
     assert form.election_rm.validators == []
 
     form = ElectionForm()
-    form.request = DummyRequest()
+    form.request = DummyRequest(session=session)
     form.request.default_locale = 'fr_CH'
     form.on_request()
     assert form.election_de.validators == []
@@ -46,7 +46,7 @@ def test_election_form_translations():
     assert form.election_rm.validators == []
 
 
-def test_election_form_model(election_day_app):
+def test_election_form_model(session):
     model = Election()
     model.title = 'Election (DE)'
     model.title_translations['de_CH'] = 'Election (DE)'
@@ -65,6 +65,7 @@ def test_election_form_model(election_day_app):
 
     form = ElectionForm()
     form.apply_model(model)
+    form.request = DummyRequest(session=session)
 
     assert form.election_de.data == 'Election (DE)'
     assert form.election_fr.data == 'Election (FR)'
@@ -111,3 +112,74 @@ def test_election_form_model(election_day_app):
     assert model.related_link == 'http://ur.l'
     assert model.tacit is True
     assert model.distinct is True
+
+
+def test_election_form_relations(session):
+    session.add(
+        Election(
+            title="First Election",
+            domain='federation',
+            date=date(2011, 1, 1),
+        )
+    )
+    session.add(
+        Election(
+            title="Second Election",
+            domain='federation',
+            date=date(2011, 1, 2),
+        )
+    )
+
+    # Add a new election with relations
+    election = Election()
+
+    form = ElectionForm()
+    form.request = DummyRequest(session=session)
+    form.on_request()
+    assert form.related_elections.choices == [
+        ('first-election', 'First Election'),
+        ('second-election', 'Second Election')
+    ]
+
+    form.election_de.data = 'Third Election'
+    form.date.data = date(2011, 1, 3)
+    form.domain.data = 'federation'
+    form.mandates.data = 1
+    form.related_elections.data = ['first-election', 'second-election']
+    form.update_model(election)
+    session.add(election)
+    session.flush()
+
+    # Change existing relations of an election
+    election = session.query(Election).filter_by(id='first-election').one()
+
+    form = ElectionForm()
+    form.request = DummyRequest(session=session)
+    form.on_request()
+    assert form.related_elections.choices == [
+        ('first-election', 'First Election'),
+        ('second-election', 'Second Election'),
+        ('third-election', 'Third Election')
+    ]
+    form.apply_model(election)
+    assert form.related_elections.data == ['third-election']
+
+    form.related_elections.data = ['second-election']
+    form.update_model(election)
+    session.add(election)
+    session.flush()
+
+    # Check all relations
+    election = session.query(Election).filter_by(id='first-election').one()
+    form.apply_model(election)
+    assert form.related_elections.data == ['second-election']
+
+    election = session.query(Election).filter_by(id='second-election').one()
+    form.apply_model(election)
+    assert set(form.related_elections.data) == set(
+        ['first-election', 'third-election']
+    )
+
+    election = session.query(Election).filter_by(id='third-election').one()
+    form.apply_model(election)
+    assert form.related_elections.data == ['second-election']